remove new http2 interface from SPI; update features
