+++ dev/com.ibm.websphere.appserver.features/visibility/private/com.ibm.websphere.appserver.internal.grpc-1.0.feature
@@ -9,6 +9,16 @@ IBM-API-Package: \
   io.grpc.stub.annotations;  type="stable", \
   io.grpc.internal; type="internal",\
 Subsystem-Name: gRPC internal 1.0
--bundles= com.ibm.ws.grpc.common.1.0
+-bundles=\
+  com.ibm.ws.app.manager.war, \
+  com.ibm.ws.grpc.common.1.0
+-features=\
+  com.ibm.websphere.appserver.artifact-1.0, \
+  com.ibm.websphere.appserver.classloading-1.0, \
+  com.ibm.websphere.appserver.appmanager-1.0, \
+  com.ibm.websphere.appserver.anno-1.0, \
+  com.ibm.websphere.appserver.containerServices-1.0,\
+  com.ibm.websphere.appserver.httptransport-1.0, \
+  com.ibm.websphere.appserver.javax.annotation-1.3; ibm.tolerates:=1.2
 kind=noship
 edition=core
