+++ dev/com.ibm.ws.transport.http/bnd.bnd
@@ -39,6 +39,7 @@ Export-Package: \
     com.ibm.ws.http.channel.outstream, \
     com.ibm.ws.http.channel.inputstream, \
     com.ibm.ws.http.logging.source, \
+    com.ibm.ws.http2, \
     com.ibm.ws.http2.upgrade
 
 Private-Package: \
@@ -87,7 +88,6 @@ Service-Component: \
     	modified:='modified';\
     	immediate:=true;\
     	properties:="service.vendor=IBM,service.ranking:Integer=0,type:String=samesite"
-    	
 
 Provide-Capability: \
     welcome.translation.provider; \
