+++ dev/com.ibm.ws.transport.http/src/com/ibm/ws/http2/Http2ConnectionHandler.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.wsspi.http.ee8;
+package com.ibm.ws.http2;
 
 import java.util.Set;
