+++ dev/com.ibm.ws.transport.http/src/com/ibm/ws/http2/Http2Stream.java
@@ -8,29 +8,39 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.wsspi.http.ee8;
+package com.ibm.ws.http2;
 
 import java.util.Map;
 
-import com.ibm.wsspi.bytebuffer.WsByteBuffer;
-
 /**
- * HTTP/2 stream handler
+ * HTTP/2 stream
  */
-public interface Http2StreamHandler {
-
-    /**
-     * Invoked when complete headers have been received for a stream
-     */
-    public void headersReady();
-
-    /**
-     * A complete body has been received for a stream
-     */
-    public void dataReady(WsByteBuffer buffer, boolean endOfStream);
+public interface Http2Stream {
 
     /**
      * Send headers to be written
      */
-    public void writeHeaders(Map<String, String> headers, boolean endOfHeaders, boolean endOfStream);
+    public void writeHeaders(Map<String, String> pseudoHeaders, Map<String, String> headers, boolean endOfHeaders, boolean endOfStream);
+
+    /**
+     * Send trailers
+     */
+    public void writeData(byte[] buffer, boolean endOfStream);
+
+    /**
+     * Send trailers
+     */
+    public void writeTrailers(Map<String, String> headers);
+
+    /**
+     * Send a RST_STREAM to terminate the stream
+     *
+     * @param reason
+     */
+    public void cancel(Exception reason, int code);
+
+    /**
+     * @return
+     */
+    public int getId();
 }
